moved calculateTimeRemaining into helpers.
